+++ Game_Dialog_MusicInfo.h
@@ -22,7 +22,7 @@ namespace Game {
 			const std::uint8_t RANEMIN;
 		public:
 			Game_Dialog_MusicInfo();
-			void getMusicInfoFromDlg(char(&filePath)[MAX_PATH],std::uint8_t& bpm, double& totalMinutes, double& beginDelay,std::uint8_t& numberOfRane);
+			void getMusicInfoFromDlg(char(&filePath)[MAX_PATH],std::uint16_t& bpm, double& totalMinutes, double& beginDelay,std::uint8_t& numberOfRane);
 		};
 	}
 }
