+++ Game_File_MusicAnalyse.cpp
@@ -2,14 +2,14 @@
 
 Game::File::Game_File_MusicAnalyse::Game_File_MusicAnalyse():COUNTOFBEAT(4){}
 
-double Game::File::Game_File_MusicAnalyse::analyseBarLengthPerMinutes(std::uint8_t& bpm) const noexcept{
+double Game::File::Game_File_MusicAnalyse::analyseBarLengthPerMinutes(std::uint16_t& bpm) const noexcept{
 	return bpm / COUNTOFBEAT;
 }
 
-void Game::File::Game_File_MusicAnalyse::analyseBarLength(std::uint8_t& bpm,double& totalMinutes, std::uint16_t& barLength) const noexcept {
+void Game::File::Game_File_MusicAnalyse::analyseBarLength(std::uint16_t& bpm,double& totalMinutes, std::uint16_t& barLength) const noexcept {
 	barLength = std::ceil(totalMinutes * analyseBarLengthPerMinutes(bpm));
 }
 
-void Game::File::Game_File_MusicAnalyse::analyse(std::uint8_t& bpm, double& totalMinutes, double& beginDelay, std::uint16_t& barLength) {
+void Game::File::Game_File_MusicAnalyse::analyse(std::uint16_t& bpm, double& totalMinutes, double& beginDelay, std::uint16_t& barLength) {
 	analyseBarLength(bpm,totalMinutes,barLength);
 }
