+++ Game_File_MusicData.cpp
@@ -1,9 +1,9 @@
 #include "Game_File_MusicData.h"
 
-Game::File::Game_File_MusicData::Game_File_MusicData(int mHandle, std::uint8_t bp, std::uint16_t bLength, double tMinutes,double bDlay,std::uint8_t numOfRane) :
+Game::File::Game_File_MusicData::Game_File_MusicData(int mHandle, std::uint16_t bp, std::uint16_t bLength, double tMinutes,double bDlay,std::uint8_t numOfRane) :
 musicHandle(mHandle),bpm(bp),barLength(bLength),totalMinutes(tMinutes),beginDelay(bDlay),numberOfRane(numOfRane){}
 
-const std::uint8_t& Game::File::Game_File_MusicData::getBpm() noexcept{
+const std::uint16_t& Game::File::Game_File_MusicData::getBpm() noexcept{
 	return bpm;
 }
