+++ Game_File_MusicFileIO.cpp
@@ -2,7 +2,7 @@
 
 std::unique_ptr<Game::File::Game_File_MusicData> Game::File::Game_File_MusicFileIO::getMusicFile() {
 	char filePath[MAX_PATH]="";//音楽ファイルのパス
-	std::uint8_t bpm = 0;//曲のbpm
+	std::uint16_t bpm = 0;//曲のbpm
 	double totalMinutes = 0;//音楽ファイルの総再生時間
 	double beginDelay = 0;//音声ファイルを再生してから曲が流れ始めるまでの時間
 	std::uint8_t numberOfRane = 4;//レーン数
